Move src/test to test
