+++ test/units/usecases/Todos.spec.ts
@@ -1,22 +1,22 @@
 import { describe, it } from "mocha";
 import { expect } from "chai";
-import { container, TYPES } from "../../../providers/container";
-import { Todos as TodoStore } from "../../../usecases/stores/Todos";
-import { Todo } from "../../../entities/Todo";
-import { Todos as UseCase } from "../../../usecases/Todos";
+import { container, TYPES } from "../../../src/providers/container";
+import { Todos as TodoStore } from "../../../src/usecases/stores/Todos";
+import { Todo } from "../../../src/entities/Todo";
+import { Todos as UseCase } from "../../../src/usecases/Todos";
 import {
   TodoCreateInput,
   TodoShowInput,
   TodoUpdateInput,
   TodoDeleteInput
-} from "../../../usecases/inputs/Todos";
+} from "../../../src/usecases/inputs/Todos";
 import {
   TodoCreateOutput,
   TodoListOutput,
   TodoShowOutput,
   TodoUpdateOutput,
   TodoDeleteOutput
-} from "../../../usecases/outputs/Todos";
+} from "../../../src/usecases/outputs/Todos";
 import { UseCaseStoreTodosMock } from "../../utils/UseCaseStoreTodosMock";
 
 describe("usecases/Todos/implementations", () => {
